Added Puzzle 1, 2, and 3 (Puzzle 3 Error)
Puzzle 1 and 2 will work. Both puzzle 2 and 3 will hide on start. Completing puzzle 1 will show puzzle 2. There is an issue with the puzzle 2 back to house button, as it doesn't allow you to go back due to a getcompnent issue. I plan to fix that thursday. Otherwise, just press F to open the Puzzlemenu. Also, there is an error when going into the Puzzle menu, but it doesn't stop anything from working.